Added perfect shot collision detection.
Perfect in the sense that it knows when a point is inside a triangle,
not perfect in the sense that it knows how big shots are drawn and
how big the outlines of the triangles are.  In that sense it's very
much imperfect, and there will likely be times in-game where someone
doesn't die when it looks like he should.  Do I care?  Not really.