Added 1/2/3 zoom keys, slowed down players more.
Now we have BZFlag-style map zoom, available on the 1, 2, and 3 keys.
This is in an effort to give players with large monitors less of an
unfair advantage.  Also, I set the player speed to half of what it was,
but this change may not be kept on a permanent basis.