Add a rotate-toggle key, draw tanks
Okay, fine, triangles really did make dodging easier than it should be
and BZFlag was a great game.  Players are tanks now.  I need to change
shot collision code in the server to reflect this, but that's fine.

You can press R to toggle whether the view rotates with you or not.  I
like having it rotate in principle, but it's a bit nauseating.